+++ packages/chakra-ui-core/src/CIcon/tests/CIcon.test.js
@@ -1,11 +1,9 @@
 import CIcon from '..'
-import { render, defaultProviders } from '@/tests/test-utils'
-import icons from '@/packages/chakra-ui-core/src/lib/internal-icons'
+import { render } from '@/tests/test-utils'
 
 const renderComponent = (props) => {
   const base = {
     components: { CIcon },
-    provide: () => defaultProviders({ $chakraIcons: { add: icons.add } }),
     template: `<CIcon name="add"></CIcon>`,
     ...props
   }