+++ packages/chakra-ui-core/src/CIconButton/CIconButton.stories.js
@@ -6,7 +6,7 @@ storiesOf('UI | IconButton', module)
     components: { CIconButton },
     template: `
       <div>
-        <CIconButton _aria-label="Phone" variant-color="blue" icon="phone" />
+        <CIconButton aria-label="Phone" variant-color="blue" icon="phone" />
       </div>
     `
   }))
@@ -14,9 +14,9 @@ storiesOf('UI | IconButton', module)
     components: { CIconButton },
     template: `
       <div>
-        <CIconButton _aria-label="Phone" size="sm" mx="3" variant-color="blue" icon="phone" />
-        <CIconButton _aria-label="Phone" size="md" mx="3" variant-color="blue" icon="phone" />
-        <CIconButton _aria-label="Phone" size="lg" mx="3" variant-color="blue" icon="phone" />
+        <CIconButton aria-label="Phone" size="sm" mx="3" variant-color="blue" icon="phone" />
+        <CIconButton aria-label="Phone" size="md" mx="3" variant-color="blue" icon="phone" />
+        <CIconButton aria-label="Phone" size="lg" mx="3" variant-color="blue" icon="phone" />
       </div>
     `
   }))
@@ -24,9 +24,9 @@ storiesOf('UI | IconButton', module)
     components: { CIconButton },
     template: `
       <div>
-        <CIconButton _aria-label="Phone" size="sm" mx="3" variant-color="green" icon="phone" />
-        <CIconButton _aria-label="Phone" size="md" mx="3" variant-color="orange" icon="star" />
-        <CIconButton _aria-label="Phone" size="lg" mx="3" variant-color="red" icon="email" />
+        <CIconButton aria-label="Phone" size="sm" mx="3" variant-color="green" icon="phone" />
+        <CIconButton aria-label="Phone" size="md" mx="3" variant-color="orange" icon="star" />
+        <CIconButton aria-label="Phone" size="lg" mx="3" variant-color="red" icon="email" />
       </div>
     `
   }))
@@ -34,7 +34,7 @@ storiesOf('UI | IconButton', module)
     components: { CIconButton },
     template: `
       <div>
-        <CIconButton _aria-label="Phone" is-loading size="lg" mx="3" variant-color="indigo" icon="phone" />
+        <CIconButton aria-label="Phone" is-loading size="lg" mx="3" variant-color="indigo" icon="phone" />
       </div>
     `
   }))
@@ -42,7 +42,7 @@ storiesOf('UI | IconButton', module)
     components: { CIconButton },
     template: `
       <div>
-        <CIconButton _aria-label="Phone" is-round size="lg" mx="3" variant-color="indigo" icon="star" />
+        <CIconButton aria-label="Phone" is-round size="lg" mx="3" variant-color="indigo" icon="star" />
       </div>
     `
   }))